refactor: API 호출 로직 분리
