--- src/page/Login/hooks/useLogin.ts
@@ -1,3 +1,0 @@
-const useLogin = () => {};
-
-export default useLogin;
